bookstore_utils-go/errors_utils/rest_errors_func.go
package errors_utils

import (
	"net/http"
)

func NewBadRequestErr(message string) *RestErr {
	return &RestErr{
		Message: message,
		Status:  http.StatusBadRequest,
		Error:   "bad_request",
	}
}

func NewNotFoundErr(message string) *RestErr {
	return &RestErr{
		Message: message,
		Status:  http.StatusNotFound,
		Error:   "not_found",
	}
}

func NewInternalServerError(message string, err error) *RestErr {
	result := &RestErr{
		Message: message,
		Status:  http.StatusInternalServerError,
		Error:   "internal_server_error",
	}
	if err != nil {
		result.Causes = append(result.Causes, err.Error())
	}
	return result
}

func NewNotAuthorizedError(message string) *RestErr {
	return &RestErr{
		Message: message,
		Status:  http.StatusUnauthorized,
		Error:   "not_authorized",
	}
}

func NewNotImplementedError() *RestErr {
	return &RestErr{
		Message: "implement me!",
		Status:  http.StatusNotImplemented,
		Error:   "not_implemented",
	}
}
